extract crapwin code for inheritance abstraction
